R6 Class computing Average Marginal Effects (AME) based on Forward Marginal Effects (FME) for a model

The AME is a simple mean FME and computed w.r.t. a feature variable and a model.

Public fields

predictor

Predictor object

features

vector of features for which AMEs should be computed

ep.method

string specifying extrapolation detection method

results

data.table with AMEs computed

computed

logical specifying if compute() has been run

Methods

Public methods

AverageMarginalEffects$new()

AverageMarginalEffects$compute()

AverageMarginalEffects$clone()

Method new()

Create a new AME object.

Usage

AverageMarginalEffects$new(
  model,
  data,
  target,
  features = NULL,
  ep.method = "none"
)

Arguments

model

The (trained) model, with the ability to predict on new data. This must be an Learner (mlr3) or train (caret) object.

data

The data used for computing AMEs, must be data.frame or data.table.

target

A string specifying the model's target variable.

features

A named character vector of the names of the feature variables for which AMEs should be computed, together with the desired step sizes.

ep.method

String specifying the method used for extrapolation detection. One of "none" or "envelope". Defaults to "none".

Returns

A new AME object.

Examples

# Train a model:

library(mlr3verse)
library(ranger)
set.seed(123)
data(bikes, package = "fmeffects")
row.id = sample(1:nrow(bikes), 100)
task = as_task_regr(x = bikes, id = "bikes", target = "count")
forest = lrn("regr.ranger")$train(task)

# Compute AMEs for all features:
overview = AverageMarginalEffects$new(
  model = forest,
  data = bikes[row.id, ],
  target = "count")$compute()
summary(overview)

# Compute AMEs for a subset of features with non-default step.sizes:
overview = AverageMarginalEffects$new(model = forest,
                                      data = bikes[row.id, ],
                                      target = "count",
                                      features = c(humidity = 0.1,
                                                   weather = c("clear", "rain")))$compute()
summary(overview)

Method compute()

Computes results, i.e., AMEs including the SD of FMEs, for an AME object.

Usage

AverageMarginalEffects$compute()

Returns

An AME object with results.

Examples

# Compute results:
overview$compute()

Method clone()

The objects of this class are cloneable with this method.

Usage

AverageMarginalEffects$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `AverageMarginalEffects$new`
## ------------------------------------------------

# Train a model:

library(mlr3verse)
library(ranger)
set.seed(123)
data(bikes, package = "fmeffects")
row.id = sample(1:nrow(bikes), 100)
task = as_task_regr(x = bikes, id = "bikes", target = "count")
forest = lrn("regr.ranger")$train(task)

# Compute AMEs for all features:
overview = AverageMarginalEffects$new(
  model = forest,
  data = bikes[row.id, ],
  target = "count")$compute()
summary(overview)
#> 
#> Model Summary Using Average Marginal Effects:
#> 
#>       Feature step.size       AME      SD SD/AME      0.25      0.75   n
#> 1      season    spring  -29.9823 31.5238   -1.1  -40.6739    -6.296  76
#> 2      season    summer    2.4663 23.3013    9.4   -8.8629   17.0032  71
#> 3      season      fall   18.8817 29.3886    1.6   -0.3961   43.4839  71
#> 4      season    winter   16.2238 23.8332    1.5    0.7489   30.5906  82
#> 5        year         1  102.2216 59.1929    0.6    33.603  152.2666  60
#> 6        year         0 -112.5327 64.6523   -0.6 -163.6236  -59.4847  40
#> 7       month         1    4.7514 11.6546    2.5   -0.6424    8.5068 100
#> 8     holiday     False    17.258    <NA>   <NA>    17.258    17.258   1
#> 9     holiday      True  -14.1095  23.564   -1.7  -29.6562   -0.3751  99
#> 10    weekday       Tue   17.8598 24.2032    1.4    1.1305   32.6675  90
#> 11    weekday       Mon   12.2788 27.6869    2.3   -7.6076   29.4393  82
#> 12    weekday       Thu    21.316 24.7356    1.2    2.6975   35.4235  83
#> 13    weekday       Sat  -53.8665 46.8255   -0.9  -79.2617  -19.0744  88
#> 14    weekday       Fri    0.5268 36.0332   68.4  -23.9033   29.4362  77
#> 15    weekday       Wed   20.4337 23.6061    1.2    1.4878   33.0241  92
#> 16    weekday       Sun  -77.5942 52.8845   -0.7 -108.7508  -31.8886  88
#> 17 workingday      True  161.1299 62.3498    0.4  122.5041  187.4092  25
#> 18 workingday     False -197.8071 91.2885   -0.5 -239.6928 -136.0076  75
#> 19    weather     clear   24.4313 39.5111    1.6    3.1349   19.8358  34
#> 20    weather     misty    0.5222 28.0608   53.7   -9.6246    0.9772  74
#> 21    weather      rain  -59.0183 52.8197   -0.9  -91.6814   -9.0111  92
#> 22       temp         1     2.508  8.5787    3.4   -1.1179    4.5872 100
#> 23   humidity      0.01    0.0116  2.2407  192.9   -0.3649    0.6915 100
#> 24  windspeed         1    -0.072  2.1247  -29.5   -0.2297      0.29 100

# Compute AMEs for a subset of features with non-default step.sizes:
overview = AverageMarginalEffects$new(model = forest,
                                      data = bikes[row.id, ],
                                      target = "count",
                                      features = c(humidity = 0.1,
                                                   weather = c("clear", "rain")))$compute()
summary(overview)
#> 
#> Model Summary Using Average Marginal Effects:
#> 
#>    Feature step.size      AME      SD SD/AME     0.25    0.75   n
#> 1  weather     clear  24.4313 39.5111    1.6   3.1349 19.8358  34
#> 2  weather      rain -59.0183 52.8197   -0.9 -91.6814 -9.0111  92
#> 3 humidity       0.1  -7.3264 15.8428   -2.2  -9.9923   1.505 100

## ------------------------------------------------
## Method `AverageMarginalEffects$compute`
## ------------------------------------------------

# Compute results:
overview$compute()