Regression data of the usage of rental bikes in Washington D.C., USA

This data set contains information on daily bike sharing usage in Washington, D.C. for the years 2011-2012. The target variable is count, the total number of bikes lent out to users at a specific day.

Usage

data(bikes)

Format

An object of class data.frame with 731 rows and 10 columns.

Source

The original data can be found on the UCI database (ID = 275).

Details

This data frame contains the following columns:

season

Season of the year

year

Year; 0=2011, 1=2012

holiday

If a day is a public holiday (y/n)

weekday

Day of the week

workingday

If a day is aworking day (y/n)

weather

Weather situation

temp

Temperature in degrees celsius

humidity

Humidity (relative)

windspeed

Windspeed in miles per hour

count

Total number of bikes lent out to users

References

Fanaee-T, Hadi, and Gama, Joao, "Event labeling combining ensemble detectors and background knowledge", Progress in Artificial Intelligence (2013): pp. 1-15, Springer Berlin Heidelberg, doi:10.1007/s13748-013-0040-3.