R6 Class computing Average Marginal Effects (AME) based on Forward Marginal Effects (FME) for a model

The AME is a simple mean FME and computed w.r.t. a feature variable and a model.

Public fields

predictor

Predictor object

features

vector of features for which AMEs should be computed

ep.method

string specifying extrapolation detection method

results

data.table with AMEs computed

computed

logical specifying if compute() has been run

Methods

Public methods

AverageMarginalEffects$new()

AverageMarginalEffects$compute()

AverageMarginalEffects$clone()

Method new()

Create a new AME object.

Usage

AverageMarginalEffects$new(model, data, features = NULL, ep.method = "none")

Arguments

model

The (trained) model, with the ability to predict on new data. This must be a train.formula (tidymodels), Learner (mlr3), train (caret), lm or glm object.

data

The data used for computing AMEs, must be data.frame or data.table.

features

A named character vector of the names of the feature variables for which AMEs should be computed, together with the desired step sizes.

ep.method

String specifying the method used for extrapolation detection. One of "none" or "envelope". Defaults to "none".

Returns

A new AME object.

Examples

# Train a model:

library(mlr3verse)
library(ranger)
set.seed(123)
data(bikes, package = "fmeffects")
task = as_task_regr(x = bikes, id = "bikes", target = "count")
forest = lrn("regr.ranger")$train(task)

# Compute AMEs for all features:
overview = AverageMarginalEffects$new(
  model = forest,
  data = bikes)$compute()
summary(overview)

# Compute AMEs for a subset of features with non-default step.sizes:
overview = AverageMarginalEffects$new(model = forest,
                                      data = bikes,
                                      features = c(humidity = 0.1,
                                                   weather = c("clear", "rain")))$compute()
summary(overview)

Method compute()

Computes results, i.e., AMEs including the SD of FMEs, for an AME object.

Usage

AverageMarginalEffects$compute()

Returns

An AME object with results.

Examples

# Compute results:
overview$compute()

Method clone()

The objects of this class are cloneable with this method.

Usage

AverageMarginalEffects$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `AverageMarginalEffects$new`
## ------------------------------------------------

# Train a model:

library(mlr3verse)
#> Loading required package: mlr3
library(ranger)
set.seed(123)
data(bikes, package = "fmeffects")
task = as_task_regr(x = bikes, id = "bikes", target = "count")
forest = lrn("regr.ranger")$train(task)

# Compute AMEs for all features:
overview = AverageMarginalEffects$new(
  model = forest,
  data = bikes)$compute()
summary(overview)
#> 
#> Model Summary Using Average Marginal Effects:
#> 
#>       Feature step.size       AME      SD      0.25      0.75   n
#> 1      season    spring   -29.472 31.5101   -39.955   -5.5139 548
#> 2      season    summer    0.4772 22.5212   -9.0235   11.6321 543
#> 3      season      fall   11.7452 28.5851   -2.4282   34.1763 539
#> 4      season    winter   15.5793 24.6394    1.6525   26.2254 551
#> 5        year         0   -99.038 67.1788 -157.0608  -20.0628 364
#> 6        year         1   97.0566  60.521   21.9401  148.0847 363
#> 7       month         1    4.0814 13.3513   -1.2566     7.459 727
#> 8     holiday     False   -1.2178 21.6103   -9.1095    9.8232  21
#> 9     holiday      True   -13.738 25.3496  -32.6323    6.2019 706
#> 10    weekday       Sat  -55.0908 49.6534  -87.6489  -15.8843 622
#> 11    weekday       Sun  -85.1527 57.7791 -122.1504  -31.8105 622
#> 12    weekday       Mon   10.7224 29.2179   -8.4101   30.4207 623
#> 13    weekday       Tue   17.9396  25.728    1.1959   32.5073 625
#> 14    weekday       Wed   20.4025 23.1599    1.3386   32.8358 623
#> 15    weekday       Thu   19.4455 24.1105   -0.3097   33.4997 624
#> 16    weekday       Fri    1.7712 35.3088  -24.8956   29.5147 623
#> 17 workingday     False -204.1875 89.3882  -257.144 -142.4332 496
#> 18 workingday      True  161.0619 62.5733  118.9398  209.6916 231
#> 19    weather     clear   26.1983 41.7886    3.5991   25.9257 284
#> 20    weather     misty     3.023 32.8661   -9.1498     0.973 513
#> 21    weather      rain  -55.3083 53.0127  -94.4096    -5.481 657
#> 22       temp         1    2.3426  7.1269   -0.4294    4.5534 727
#> 23   humidity      0.01   -0.2749   2.626   -0.3249    0.3504 727
#> 24  windspeed         1    0.0052  2.4318   -0.1823    0.2318 727

# Compute AMEs for a subset of features with non-default step.sizes:
overview = AverageMarginalEffects$new(model = forest,
                                      data = bikes,
                                      features = c(humidity = 0.1,
                                                   weather = c("clear", "rain")))$compute()
summary(overview)
#> 
#> Model Summary Using Average Marginal Effects:
#> 
#>    Feature step.size      AME       SD      0.25     0.75   n
#> 1  weather     clear 26.19828 41.78861   3.59911 25.92567 284
#> 2  weather      rain -55.3083 53.01272 -94.40958 -5.48097 657
#> 3 humidity       0.1 -7.38328 16.35698   -8.5513  1.07549 727

## ------------------------------------------------
## Method `AverageMarginalEffects$compute`
## ------------------------------------------------

# Compute results:
overview$compute()